batch: enable task routes for frontend and batch when ENABLE_BATCH=True
To enable the batch test prior to this change, it was necessary to set
`ENABLE_BATCH = True` and to remove the `CELERY_TASK_ROUTES` for the
frontend. This commit simply enables both when `ENABLE_BATCH = True`.

The effect of this change is that frontend tasks are routed to the same
celery queues regardless of whether the batch testing facility is
enabled. This simplifies understanding of the system.

This change necessitates a configuration change to the way Celery is
started. Assign workers to the `db_worker` and `slow_db_worker` queues.

Signed-off-by: Maarten Aertsen <maarten.aertsen@ncsc.nl>