batch: define non-overlapping celery queues for batch test
This commit changes queue configuration in the following way:
  - move the scheduler to its own queue `batch_scheduler`
  - rename `worker_slow` to `batch_slow` for the batch_async* tasks

This change necessitates a configuration change to the way Celery is
started. Assign workers to the two queues defined above.

To retain the previous mapping between queues and workers,
explicitly map the `batch_scheduler` and `worker_slow` to the worker
that previously implicitly worked the default the `celery` queue.

Signed-off-by: Maarten Aertsen <maarten.aertsen@ncsc.nl>