Fix #695 - Remove the use of pip-sync
As explained in #695:
The purpose of pip-sync is to guarantee that a local environment is exactly
per the requirements. With our complex dependencies, this purpose can already
not be met. Working around it seems like we adopt a tool to then only sabotage
its core function. Note that make venv already uses pip install rather than
pip-sync, which I think is the right approach with our current
dependency situations.