Ref #733 - Run CI on main+release push - may fix leftover cache issues
According to https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows#matching-a-cache-key
GitHub only looks for a cache hit in the current (feature) branch, and then in the parent.
Since we don't build on main, there is never a cache hit in new feature branches.

We should probably be building on main anyways, and this may finally fix cache misses.