Fix #769 - Correctly indicate probe error to user (#802)
Before this change, the only check was whether a test was complete,
not whether it was at all successful. After redirect to the result
page, we then presented the most recent result for each probe,
which could silently return results from the distant past.

There may still be some cases where test error handling is not optimal,
but those should be loud, which means #770 will notify us.