Allow forked PR's (eg: dependabot) to run builds
This uses a different method for sharing the container images between jobs depending if the PR if from a fork or not. Forked PR's don't have access to the registry (rightfully so), but using artifacts works. But this is really slow. So for non-forked PR's we use the registry (this also allows to deploy the branch on a development/staging machine).