Ref #1130 - Partial revert of c5ba541 / #904 to allow cert detection in bad ssl
While using a standard HTTP client is good, it does mean we
can't connect to some very obscure setups with it anymore.
In the case of cert detection, the HTTP client was exclusively
used for guessing if there even is any SSL, while the rest of the
code can still handle very bad configs.

As we needed our legacy clients in a few places anyways until
finishing #714, this commit reverts the cert check back to
the legacy client, allowing cert checks in these very bad configs.