[1.8.x] Remove docker image branch tags (#1189)
This was a blocker for builds on branches with slashes. We also don't need it that much, as branch following will be implemented in #1155 by fetching the branch sha and using that for image pulls.

(cherry picked from commit de7e7f1c2edcb5164050bf4ed7fc9a213242ce9a)