Fix #1540 - Make RPKI probe verdict SUCCESS even if some tests are skipped
This happens when e.g. there are no mail servers to test. By default,
because worst_status is FAIL, this would be counted as a FAIL for the

Ref #1003
probe verdict (shown at the top of the page).
RPKI has an exception case now, where it can be worst_status=FAIL
and still be partially optional.