Get rid of almost all static IP configuration
To allow multiple instances to be easily configured to run next to each
other on the same server we want to reduce the amount of static IP
configuration of containers.

- Move static files used for testing (hof, public_suffix_list) into
separate service to keep it distinct from test-target

- remove static IP's for almost all containers/services. Only a few
static IP's are kept, mostly for dev/test environments where these IP's
are referenced in the zone files for resolving by the mock-resolver.

- configure nginx to do DNS lookups for proxy_pass's to backend
containers for app and monitoring.

- no longer configure the internal resolver (resolver-validating) as
container resolver. Since the application configures a resolver for use
internally which explicitly uses the `resolver-validating` this
configuration is not needed, it also enables resolving containers by
name.

- configure the `app` resolver to lookup the `resolver-validating`
containers IP by container name.

- restart services that use `resolver-validating` when that services is
restarted (and potentially get's a different IP).